backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=3.14
djangorestframework-simplejwt>=5.2
celery>=5.2
redis>=4.5
channels>=4.0
channels-redis>=4.0
python-dotenv>=1.0
gunicorn>=20.1
uvicorn>=0.22.0
openai>=1.0.0

